chore: refactor reference-style owner-auth schema and operation implementations
