+++ packages/amplify-graphql-api-construct-tests/src/__tests__/graphql-schemas/reference-style-owner-auth/mutations.ts
@@ -1,7 +1,7 @@
 /* tslint:disable */
 /* eslint-disable */
 
-import * as APITypes from '../API';
+import * as APITypes from './API';
 type GeneratedMutation<InputType, OutputType> = string & {
   __generatedMutationInput: InputType;
   __generatedMutationOutput: OutputType;
