+++ packages/amplify-graphql-api-construct-tests/src/__tests__/graphql-schemas/reference-style-owner-auth/queries.ts
@@ -1,7 +1,7 @@
 /* tslint:disable */
 /* eslint-disable */
 
-import * as APITypes from '../API';
+import * as APITypes from './API';
 type GeneratedQuery<InputType, OutputType> = string & {
   __generatedQueryInput: InputType;
   __generatedQueryOutput: OutputType;
