+++ packages/amplify-graphql-api-construct-tests/src/__tests__/graphql-schemas/reference-style-owner-auth/schema-primary.graphql
