+++ packages/amplify-graphql-api-construct-tests/src/__tests__/owner-auth/assoc-field/test-implementations.ts
@@ -1,30 +1,17 @@
-// NOTE: The API and operations used in this test were generated by running `amplify codegen --maxDepth 4`, then edited to remove createdAt,
-// updatedAt, and __typename.
-
-import { AppSyncGraphqlResponse, doAppSyncGraphqlMutation, doAppSyncGraphqlQuery } from '../../../utils';
 import {
-  CreatePrimaryMutation,
-  CreateRelatedManyMutation,
-  CreateRelatedOneMutation,
-  GetPrimaryQuery,
-  GetRelatedManyQuery,
-  GetRelatedOneQuery,
-  ListPrimariesQuery,
-  ListRelatedManiesQuery,
-  ListRelatedOnesQuery,
-  UpdatePrimaryMutation,
-  UpdateRelatedManyMutation,
-  UpdateRelatedOneMutation,
-} from './API';
-import {
-  createPrimary,
-  createRelatedMany,
-  createRelatedOne,
-  updatePrimary,
-  updateRelatedMany,
-  updateRelatedOne,
-} from './graphql/mutations';
-import { getPrimary, getRelatedMany, getRelatedOne, listPrimaries, listRelatedManies, listRelatedOnes } from './graphql/queries';
+  doCreateRelatedOne,
+  doCreateRelatedMany,
+  doCreatePrimary,
+  doUpdatePrimary,
+  doGetPrimary,
+  doListPrimaries,
+  doUpdateRelatedOne,
+  doGetRelatedOne,
+  doListRelatedOnes,
+  doUpdateRelatedMany,
+  doGetRelatedMany,
+  doListRelatedManies,
+} from '../../graphql-schemas/reference-style-owner-auth/operation-implementations';
 
 // #region Primary as source
 
@@ -599,197 +586,6 @@ export const testListRelatedManiesRedactedForDifferentOwners = async (
   expect(relatedMany).toBeDefined();
   expect(relatedMany.id).toEqual(relatedManyId);
   expect(relatedMany.primary).toBeNull();
-};
-// #endregion
-
-// #region Implementations
-
-const doCreateRelatedOne = async (
-  apiEndpoint: string,
-  accessToken: string,
-  relatedOneId: string,
-  primaryId: string,
-): Promise<AppSyncGraphqlResponse<CreateRelatedOneMutation>> => {
-  return doAppSyncGraphqlMutation({
-    apiEndpoint,
-    auth: { accessToken: accessToken },
-    query: createRelatedOne,
-    variables: {
-      id: relatedOneId,
-      primaryId,
-    },
-  });
-};
-
-const doCreateRelatedMany = async (
-  apiEndpoint: string,
-  accessToken: string,
-  relatedManyId: string,
-  primaryId: string,
-): Promise<AppSyncGraphqlResponse<CreateRelatedManyMutation>> => {
-  return doAppSyncGraphqlMutation({
-    apiEndpoint,
-    auth: { accessToken: accessToken },
-    query: createRelatedMany,
-    variables: {
-      id: relatedManyId,
-      primaryId,
-    },
-  });
-};
-
-const doCreatePrimary = async (
-  apiEndpoint: string,
-  accessToken: string,
-  primaryId: string,
-): Promise<AppSyncGraphqlResponse<CreatePrimaryMutation>> => {
-  return doAppSyncGraphqlMutation({
-    apiEndpoint,
-    auth: { accessToken: accessToken },
-    query: createPrimary,
-    variables: {
-      id: primaryId,
-    },
-  });
-};
-
-const doUpdatePrimary = async (
-  apiEndpoint: string,
-  accessToken: string,
-  primaryId: string,
-): Promise<AppSyncGraphqlResponse<UpdatePrimaryMutation>> => {
-  return doAppSyncGraphqlMutation({
-    apiEndpoint,
-    auth: { accessToken: accessToken },
-    query: updatePrimary,
-    variables: {
-      id: primaryId,
-    },
-  });
-};
-
-const doGetPrimary = async (
-  apiEndpoint: string,
-  accessToken: string,
-  primaryId: string,
-): Promise<AppSyncGraphqlResponse<GetPrimaryQuery>> => {
-  return doAppSyncGraphqlQuery({
-    apiEndpoint,
-    auth: { accessToken },
-    query: getPrimary,
-    variables: {
-      id: primaryId,
-    },
-  });
-};
-
-const doListPrimaries = async (
-  apiEndpoint: string,
-  accessToken: string,
-  primaryId: string,
-): Promise<AppSyncGraphqlResponse<ListPrimariesQuery>> => {
-  return doAppSyncGraphqlQuery({
-    apiEndpoint,
-    auth: { accessToken },
-    query: listPrimaries,
-    variables: {
-      id: primaryId,
-    },
-  });
-};
-
-const doUpdateRelatedOne = async (
-  apiEndpoint: string,
-  accessToken: string,
-  relatedOneId: string,
-  primaryId: string,
-): Promise<AppSyncGraphqlResponse<UpdateRelatedOneMutation>> => {
-  return doAppSyncGraphqlMutation({
-    apiEndpoint,
-    auth: { accessToken },
-    query: updateRelatedOne,
-    variables: {
-      id: relatedOneId,
-      primaryId,
-    },
-  });
-};
-
-const doGetRelatedOne = async (
-  apiEndpoint: string,
-  accessToken: string,
-  relatedOneId: string,
-): Promise<AppSyncGraphqlResponse<GetRelatedOneQuery>> => {
-  return doAppSyncGraphqlQuery({
-    apiEndpoint,
-    auth: { accessToken },
-    query: getRelatedOne,
-    variables: {
-      id: relatedOneId,
-    },
-  });
-};
-
-const doListRelatedOnes = async (
-  apiEndpoint: string,
-  accessToken: string,
-  relatedOneId: string,
-): Promise<AppSyncGraphqlResponse<ListRelatedOnesQuery>> => {
-  return doAppSyncGraphqlQuery({
-    apiEndpoint,
-    auth: { accessToken },
-    query: listRelatedOnes,
-    variables: {
-      id: relatedOneId,
-    },
-  });
-};
-
-const doUpdateRelatedMany = async (
-  apiEndpoint: string,
-  accessToken: string,
-  relatedManyId: string,
-  primaryId: string,
-): Promise<AppSyncGraphqlResponse<UpdateRelatedManyMutation>> => {
-  return doAppSyncGraphqlMutation({
-    apiEndpoint,
-    auth: { accessToken },
-    query: updateRelatedMany,
-    variables: {
-      id: relatedManyId,
-      primaryId,
-    },
-  });
-};
-
-const doGetRelatedMany = async (
-  apiEndpoint: string,
-  accessToken: string,
-  relatedManyId: string,
-): Promise<AppSyncGraphqlResponse<GetRelatedManyQuery>> => {
-  return doAppSyncGraphqlQuery({
-    apiEndpoint,
-    auth: { accessToken },
-    query: getRelatedMany,
-    variables: {
-      id: relatedManyId,
-    },
-  });
-};
-
-const doListRelatedManies = async (
-  apiEndpoint: string,
-  accessToken: string,
-  relatedManyId: string,
-): Promise<AppSyncGraphqlResponse<ListRelatedManiesQuery>> => {
-  return doAppSyncGraphqlQuery({
-    apiEndpoint,
-    auth: { accessToken },
-    query: listRelatedManies,
-    variables: {
-      id: relatedManyId,
-    },
-  });
 };
 
 // #endregion
